pyproject.toml
[project]
name = "ai-recipe-planner"
version = "0.1.0"
description = "AI-powered meal planning and recipe management application"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.40.0",
    "streamlit>=1.41.0",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.0",
    "streamlit-authenticator>=0.4.2",
    "pyyaml>=6.0.3",
    "bcrypt>=5.0.0",
    "ruff>=0.14.7",
    "google-genai>=1.56.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "types-Pillow>=10.2.0",
    "pre-commit>=4.0.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["lib"]


[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "streamlit.*"
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "E402",  # module level import not at top of file (needed for dotenv/streamlit)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=lib",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["lib"]
omit = ["tests/*", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
]
